Add support for exclusionary text filters
Right now the way to create an exclusionary text filter is to prefix the name with ‘!’ character.
Any exclusion filter that matches will prevent a given log line from appearing.
I can see a future change being to add a control in the text filter creation editor to mark something as “exclusionary” or not.